MXS-4710 Add NEED_DNS server status bit
Because name lookups are run in threadpool when MaxScale starts,
the servers will not have binary addresses available immediately.
Since the lookup can sometimes take a while, its status needs to
be tracked so that pending servers are not used for backend
connections. This only affects MaxScale availability right after
startup.