MXS-4866 Remove `overridingModuleParams` props
The mentioned props is used in the settings page to handle
the `log_throttling` parameter. The default value of that
parameter is an object which is a parameter with nested
parameters. The props is used to break that nested parameters
into flat parameters. This will be also done for the `connection_metadata`.

This makes `moduleParameters` as a required props.