MXS-4869 Remove `pauseCompute` props of `data-table`
The props was introduced to prevent the table data from
being updated while the table is in edit mode since
the introduction of the refresh rate feature. However, it
actually does more harm than good. After updating
the parameter of an object, the table doesn't render with
new data because the watcher is for the `data` props,
not the `pauseCompute` props.