Schedule server DNS lookups manually
This prevents the problem where the deletion of the server caused the
scheduled threadpool task to access freed memory when the server was
deleted in the same stack frame. By starting the DNS lookup task
manually once it's known that the server is valid, the problem goes away
as it won't be deleted after it's been added to the global server list.