MXS-4879: Fix transaction tracking in 23.08
The transaction tracking in the client protocol was not done if the
router did not explicitly request it. This caused MXS-4879 to also
affect readwritesplit whereas in older versions it only affected
readconnroute.

As the functionality is needed by `maxctrl reload session` and
idle_session_pool_time, both of which are core features, the tracking
must be done unconditionally.

This commit also fixes the problem where the
QueryClassifier::update_from_reply() call was unconditionally made for
all completed requests. It should only be done when the
RCAP_TYPE_SESCMD_HISTORY capability bit is set.