MXS-4840: Move sescmd history statistics into the core
The feature is a core part of MaxScale and the statistics are now there
as well.

The average length is now an exponentially moving average with the same
alpha values as the ones in the server response times in the
mxs::Target. A moving average is also easier to implement atomically as
it only depends on the previous value and a compile time constant.