MXS-4791 Remove update ordering from the Collector
The feature's original implementation had a flaw in that it
could cause exessive memory use. The fix to that flaw caused
the feature to give even less predictability than before.

The ordering was based on a timestamp generated at the point
when the event was added to the ShareData queue. Because the
insert to the queue could be delayed in the case where the
queue was full, the timestamp generation could be delayed
compared to other threads. If the consumer of events
(a subclass of the Collector) needs absolute ordering of events,
it will have to do so itself. The events generated by one thread
are always processed in order.