MXS-4896 Handle eviction more efficiently
Earlier during eviction we picked by random a hash bucket and
if it was empty, we picked by random a new hash bucket until
a non-empty one was found.

Now the starting point is picked randomly and if it happens to
be empty, we linearly loop over the collection until a non-empty
one is found.