MXS-4896 Handle query classifier cache size change correctly
If it is detected that the size of the query classifier size
has changed, an lcall() where the size adjustment is made is
scheduled.

Adjusting the cache size while the cache is being used turned
out to be quite tricky. The size of the cached data will change
if only a restricted set of information is collected the first
time a statement is parsed. However, the total size of the cache
must be updated only if the statement is still in the cache and
that turned out to be quite tricky to figure out without making
an extra cache lookup.