MXS-4907: Fix nested parameters in mxs::Config
The parameters were only added in the JSON output that was used in the
REST-API. They also need to be added whenever the configuration
parameters are being modified at runtime.

Added a test case that verifies the problem in 23.02. In 22.08 the
problem does not appear with the test which suggests that the
requirements of nested parameters might not be validated correctly.