Fix idle_pool_max_conns
The extra connections need to be closed to avoid having too many
connections at the same time.

The parallel querying part of the test also now correctly checks that
the queries are actually sent and received with no errors.

Lowered the value of idle_session_pool_time to 50ms to make the test
faster. This way the first part of the test completes much quicker.