MXS-4910: Remove unconditional query classification
The query classification costs about 15% in terms of raw query execution
time when it comes to database dumps. Since the reason for the parsing
was to track query latencies, it should not be done unless it costs
nothing.

If the statistics are deemed important enough or a user wants to make
them accurate regardless of the cost, a generic service option that
allows extra capability bits to be injected could be added to force
query classification even on services that don't need it.