Disable master_reconnection if sescmd history is off
This commit changes the behavior of readwritesplit so that an old
configuration that had `disable_sescmd_history=true` will still work
even with the change of the default values. Without it, all upgrades
from older versions that had the session command history disabled would
cease to function.

Documented the change in behavior in the readwritesplit documentation.