Simplify query routing in schemarouter
Moved the shard map interaction and session initialization from the main
routeQuery function into smaller ones. This reduces the size of the
function which should make it easier to comprehend.

Replaced the error packets that were sent to clients with calls to
Session::kill(). This simplifies the code by removing the need to return
error values and also makes the fact that the session was killed
clearer.