Don't retry failed routeQuery calls in readwritesplit
Currently if they are retried, there's a possibility that the component
below the router is not aware that the query was aborted and retried
which can lead to bookkeeping errors.

The only way to handle this is for the ServiceEndpoint to inspect the
return value and call the handleError method of the parent component to
resolve the error. This way all errors are handled in the same place and
no modifications to modules are needed.