MXS-4783: Allow exceptions from all API functions
The clientReply and handleError API functions can now also throw
exceptions.

If clientReply throws an exception, the session will be closed; this is
the same as returning false would do. The return value in clientReply is
largely useless as routing replies doesn't seem to have any error
conditions in the current codebase.

If handleError throws an exception, it is propagated upwards to the
parent component. In case of a single-level service, this is the
MXS_SESSION which will simply kill the session.

The explicit closing of downstream Endpoints that the ServiceEndpoint
did must not be done as the router session will also do it. Since the
router sessions use the Backend type to deal with the Endpoints, their
states will not match if the Endpoint is closed by anything other than
the router itself.

Documented the exception behaviors of all API functions into the
relevant locations and added extra comments to clarify why calls to
mxs::Component::handleError cannot throw but calls to
mxs::Routable::handleError may throw.