MXS-4883: Add resultset metadata to the JSON
The information was readily available and can be sent alongside the
results. To keep the API backwards compatible, the `fields` field was
left as-is and the information was added to the `metadata` field.

Added a test case that uses a few data types in a table to make sure
that each field gets processed correctly.