MXS-4885: Add simple filtering expressions
The simple expressions are mostly for (non)equality and ordering. This
alone won't be that useful so at least AND and OR expressions are
needed.

In addition to the expressions, either a from-date expression is needed
or the logging output must include a UNIX timestamp that can be compared
as an integer.