MXS-4885: Add /maxscale/logs/entries endpoint
Since the /maxscale/logs/data endpoint does not return a resource
collection, it cannot be filtered using the `filter` request option. To
make it, a new endpoint was needed. The /maxscale/logs/entries endpoint
is nearly identical to the old endpoint except that each row is its own
resource.

Unlike a normal resource collection, this one does not have an endpoint
for accessing individual values. Given that this use-case isn't really
that sensible, it was left out.

Added schema validation tests for both the new and the old endpoints.

Documented the new endpoint and regenerated the relevant documentation.