Move PS handling into QueryClassifier
By classifying a COM_STMT_CLOSE also as mxs::sql::TYPE_DEALLOC_PREPARE,
the need to look at the command byte goes away. This makes it possible
to do the PS handling completely inside the QueryClassifier.

Making the updating of the RouteInfo a two step process allows the delta
list for the temporary tables to be removed. This reduces the amount of
work that is done by postponing the temporary table handling until the
commit step at which point there is no need to keep the dropped or added
tables in memory.

The QueryClassifier::update_from_reply() function now also tracks the
parameter counts for the prepared statements. The
mxs::Reply::generated_id() is the same ID that is used to store the
prepared statement. By detecting positive the parameter counts, it is
possible to also move this into the QueryClassifier, removing the need
for QueryClassifier::ps_store_response().