Use delay_routing() in the cache filter
If the mxs::Component to which the filter belongs to would be closed by
an event that gets processed before the lcall() is processed, the filter
would access potentially freed memory. Using delay_routing() both avoids
this problem and also correctly propagates routing failures to the
parent component.

Also adjusted delay_routing() to use a lcall() instead of a dcall() in
the case where the delay was set to zero. There should be no real
difference between a one millisecond delay and no delay at all. Also
removed the unused m_worker member from MXS_SESSION.