MXS-4837: Allow `targets` in `alter`
The REST-API itself doesn't allow mixed types in the API which means
only servers must be in the `servers` relationship and only services in
the `services` relationship. Modifying the `targets` of a service will
thus require an extra GET request that is then used to determine whether
the target is a server or a service.