MXS-4932: Remove gwbuf_append()
No longer needed as it can be replaced with GWBUF::append(). The only
use-cases where it was used were also situations where the buffer was
always not null and thus the checks done by gwbuf_append() weren't
necessary.

This also fixes the stale pointer in the cache filter that was passed
around but was not used for anything.