MXS-4932: Output GWBUF values from get_value()
The only notable functional change was the removal of the pResponse
GWBUF* argument from get_value_handler(). The buffer is now freed
automatically when it goes out of scope.

The remainder of the changes were mostly mechanical substitutions of
GWBUF pointers to GWBUF values and the removal of gwbuf_free() calls.