Fix test_small_vector build failures on Ubuntu Jammy
For whatever reason, the template code in test_small_vector ends up
causing a -Wstringop-overflow warning when the type that is used is
mxb::small_vector<std::string, 2>.

The warning originates from the __gnu_cxx::__aligned_membuf storage
that's used by the variant type to store the vector that holds the data
when it's moved onto the heap. The line of code that triggers it is the
call into std::vector<T>::erase(). If std::get_if is used instead of
std::get, the warning goes away. Since this is only for this one test
and the code appears to work as expected (no ASAN or UBSAN complaints),
it's better to disable the warning for the test than to modify the code.