MXS-4932: Remove buffer.hh free functions
Replaced the use of gwbuf_free() with delete statements. There was only
three places where it was still used.

Moved the GWBUF <-> GWBUF* conversion functions into the nosql namespace
to make it clear that they're only used there. Added a TODO note that
mentions the Jira task (MXS-4931) for converting the nosqlprotocol to
use GWBUF values.