MXS-4466 Handle deleting and adding annotation lines
The users can now delete existing annotation lines or
adding new ones.

This also removes the default annotation lines of dashboard graphs
and the `toggle` button to set annotation visibility as they
would increase the complexity in the user interface.

The chart options reactivity issue is also fixed in
`mxs-line-chart-stream` component which happens when the
annotation line is deleted.