MXS-4466 Improve UX for add button
Having a big rectangular button for adding annotation object
when there is no annotation object is nice. However, when there
are already some annotation objects, the users may accidentally
click on the button after changing the color using the color picker.

This shows the big button only when there is no annotation object
and shows a smaller button on the right side of the heading section.

This also fixes `onSave` method as it didn't update the change of
inactive graph configs.