MXS-4921 & MXS-4922: Remove COM_STMT_CLOSE assertion
If a COM_STMT_CLOSE is received that doesn't have a COM_STMT_PREPARE, it
is not possible to know whether it is a truly unknown COM_STMT_CLOSE or
whether the COM_STMT_CLOSE was for a COM_STMT_PREPARE that was closed
while the backend connection was being opened.

This is the assertion that caused the mxs3499_causal_reads_ps test to
fail.