MXS-4937: Compare get_canonical implementations
Added a new test case that compares the output of the generic and the
specialized get_canonical() functions. This guarantees that any
modifications to the code do not cause the results to differ from the
generic implementation that is now used as the baseline value.

The only difference in the output produced by the AVX2 implementation
and the generic one was that the generic version removed trailing
whitespace. The amount of trailing whitespace is not expected to change
wildly in production usage and as the de facto standard implementation
before MaxScale 24.02 has been the AVX2 one, the code in the generic
version was removed instead of it being added to the AVX2
implementation.