MXS-4937: Move marker processing out of simd256
The only part that needs the specialized implementation is the creation
of the markers. The processing of the markers is done one element at a
time and it doesn't really benefit from being compiled as an AVX2
capable function.

This is the first step in making the generic code use the same marker
creation code that the AVX2 implementation uses.