MXS-3620 Fix custom range when the end of time range is today
When using `v-date-picker`, it returns an ISO date string.
So when the `to` value is set to today, the `strToTimestamp`
method returns the timestamp starting at the beginning of the
day instead of the current timestamp.

This also uses dynamic width, so when using a predefined range,
it has a compact width and expands when a custom range is chosen.