MXS-3620 Handle `NOW` value of TIME_REF_POINTS better
If the value is `NOW`, it's not actually the current timestamp but
the value at the time when the `date-range-picker` component
updates its model value. As a result, it doesn't guarantee the
correct timestamp.

This refactors the `date-range-picker` component to return
`TIME_REF_POINTS` or ISO date strings. This way, when users click
the Filter button, the `date_range` values can be parsed into
timestamps accurately.