MXS-4937: Add generic marker generation function
The function is used when the CPU does not support the AVX2 instruction
set. In practice this is only used by ARM CPUs as most CPUs nowadays
support AVX2.

The function performs about as well as the original one except when the
input consists largely of numbers. In this case it performs worse as the
marker processing seems to add an overhead that the old generic version
did not have. Further optimizations to the number parsing are likely to
improve the performance. For inputs that consist only of SQL keywords,
the new generic versions seems to outperform the old one by about 30%
but this result is somewhat artificial.

The old generic version is exposed for the sake of testing. This allows
it to be used as the verification function as it produces a stable
result.