MXS-4937: Add number skipping optimization
The probing of numbers done by process_markers() was one of the slowest
parts of it. This was especially true when large IN() lists were being
parsed. The worst-case scenario of a large IN list that consists of only
one digit was where the SIMD version was about 50% slower than the
scalar version. In the general case where no repeating numbers were
present, the SIMD version was still slower but not as dramatically slow.
For the usual case where numbers do not repeat, the SIMD version still
fell behind by roughly 15% when compared to the generic scalar version.

For comparison, the case where there are no digits and only identifiers,
the SIMD version is over three times faster than the scalar version.

By skipping of repeated digits and dots whenever a digit marker is read,
the performance of the SIMD version is on par with the scalar versions
or at least improves when compared to not skipping digits.

The old generic version was also modified to skip over repeating digits
and dots. This makes it so that the output of all the versions produces
similar output even with malformed outputs or outputs that aren't normal
SQL statements. A good example of this kind of output is the `FOR 1..10`
SQL expression. The old generic version converted this into `FOR ?.?`
and the new ones converted it into `FOR ?`. The latter is preferred in
all cases.