MXS-4937: Add argument extraction
The "negative" of the canonical can now be extracted with the
maxsimd::get_canonical_args() function.

The added test case verifies that both the generic and the specialized
implementation produce the same canonical form and the same
arguments. Since it's the same code that does the normal canonical query
transformation, this is somewhat redundant but it's better to test it
too much than not enough.