MXS-3620 Extend `mxs-filter-list` to has a customized appearance
To highlight the Filter button, the activator of the `mxs-filter-list>`
component needs to have a different appearance.
The style should be similar to the `<v-select>` dropdown style,
which changes its color only when the menu dropdown is opened.

This adds `changeColorOnActive` props to enable this style.
`activatorProps` is also added to futher customize the appearance
of the button.