MXS-3620 Add `reverse` props to `mxs-filter-list`
This component was refactored in MXS-4728 to return unselected items,
which helps to reduce computation steps for the `mxs-virtual-table`
component. However, from the filtering UX perspective, not all item
need to be selected by default. Only when the `Select All` checkbox
is present, the selection of all items is expected.

This adds the `reverse` prop to control this behavior. Additionally
this also adds the `hideSelectAll` prop to conditionally hide the
`Select All` checkbox.