MXS-4937: Recombine canonical and arguments to SQL
This is the inverse of get_canonical_args() and makes a round-trip via
canonical_args_to_sql() possible.

The unit test checks that both the generic and the specialized versions
produce the same result. This is again somewhat redundant as the
canonical form calculation is already tested but if there's ever
specialized version of recombination (unlikely), it'll get tested
automatically.