Fix JSON after creating service
The injection of the relationships as parameters modified the original
JSON into a form that would be rejected if given to
config_runtime_alter_service_from_json(). The configuration
synchronization needs the JSON to remain valid and thus the
modifications must be reverted.

This fixes the config_sync test that was broken by commit
295bbb36a277f68d6d9ddf3cc714f6e147f44dcc.