MXS-4945 Address the issue with object name uniqueness validation
The uniqueness of the object name is not limited to a specific type
but applies across all types. For example: If there is already a
monitor name `Monitor`, creating a new server named `Monitor` would
receive an error as follows:
`Can't create server 'Monitor', a monitor with that name already exists`

This adds `fetchAllMxsObjIds` action to fetch all object ids and uses
it for validating object name uniqueness.