MXS-4947: Use default db for "universal" tables
If all the tables or databases in a query are present on all the nodes,
the default database is used to restrict it to a subset of these
nodes. The intention of this is to allow the default database to fully
dictate where queries that otherwise could be routed anywhere are to be
routed. This was already possible with queries that do not use a table
(e.g. SHOW DATABASES) but for queries with these sort of "universal"
tables like `information_schema` were not routed the way they were
supposed to be.

Documented the behavior and extended the schemar_implicit_db test case
to cover this bug.