MXS-4936: Colorize output in compare_canonicals
If the define is changed from 0 to 1, the output is displayed in
inverted video coloring. This makes it easier to see differences in
trailing whitespace which is helpful when comparing the different
implementations. The default is to not use it since there's no detection
whether the terminal supports colors or not.

Also added a counter for errors that makes it easier to see if new
errors were added or removed when modifications are made.