MXS-4936: Ignore backslash-escaped markers
This way the query `SELECT \0` is handled the same way by both the old
generic version and the new generic and SIMD versions. The old generic
version correctly converted it to not contain a question mark whereas
the new ones added a backslash and a question mark.

With this, all of the canonical query generation functions produce the
same output with all the test files.