MXS-4940: Add mariadb::PsToText class
The class is similar to the mariadb::QueryClassifier in that the
track_query() and track_reply() functions must be called in the
routeQuery() and clientReply() functions.

In track_query() a reference to the original COM_STMT_PREPARE is stored
until it is known that the execution succeeds. Once it does and
track_reply() is called, the SQL of the prepare and the parameter count
are stored. In addition to this, the only thing that is needed is to
extract the parameter types from the COM_STMT_EXECUTE and use the types
to decode the binary form into text.

The actual conversion from the binary from to text form is not yet
implemented.