MXS-4940: Add skeleton for parameter conversion
The skeleton code extracts the required parts from the COM_STMT_EXECUTE
and passes them onto the function responsible for decoding the binary
values into text. The actual decoding of the binary values is not yet
implemented.

The type information in the COM_STMT_EXECUTE is only sent for the first
execution of a COM_STMT_EXECUTE which needs to be kept in memory for all
subsequent executions.