MXS-4940: Decode binary types into text
The integer, floating point and string types are relatively simple to
handle. The time and timestamp types are a bit trickier to deal with as
they are encoded in a variable sized struct and must be converted into
the correct string format.

The values returned by the binary_to_text() function are suitably
escaped for execution as SQL.