MXS-4959: Log COM_STMT_EXECUTE as SQL in qlafilter
With the mariadb::PsToText class, the COM_STMT_EXECUTE can now be
converted into its text form with relative ease. This makes it possible
to use the qlafilter as a general SQL log where previously it was unable
to capture the actual executed SQL if it was using the binary protocol.