MXS-4921: Fix debug assert in read_history_response()
If a COM_STMT_PREPARE is being executed by a backend connection and the
client closes it with a COM_STMT_CLOSE, the response to it will be
removed before the history replay is complete. This caused a debug
assertion in read_history_response() to be hit when the response to it
is not found. This was visible as a debug assertion that was hit by the
mxs2295_change_user_loop test.

Storing the expected responses to the session command history at an
earlier point in send_history() solves the problem. The storing of the
IDs is now redundant but they're still kept for the log_info messages.