MXS-4649 Unifying common helper methods into one module
Separating common workspace helpers, shared common helpers, and
MaxGUI app helpers saves a few hundred kilobytes in the build, which
is nice but not a significant amount.

Having just one helpers module registered globally is better for
maintenance and ease of work. It also prevents unexpected overlapping
of function names.