MXS-4961: Send last error message in kill()
This makes it so that a KILL CONNECTION_ID() will appear identical both
with a direct connection and with a connection through MaxScale.

Fixed the SQLSTATE for the ER_CONNECTION_KILLED error that was generated
by the client protocol.

Added a test case that verifies that the behavior is the same as the one
in MariaDB server.