MXS-4959: Track large packets in qlafilter
The new MultiPartTracker class implements the similar packet boundary
tracking functionality that is already implemented in multiple
places. This makes the component reusable and the first use-case for it
is in the PsTracker (the old PsToText class) where it can be used to
automatically ignore trailing parts of large queries and LOAD DATA LOCAL
INFILE execution.

Renamed the files from ps_to_text to trackers as there are now multiple
tracker classes in there.

Added a limitations section to the qlafilter documentation and explained
the behavior of SQL queries that span multiple packets.