Don't call routeQuery from clientReply
This will cause problems if the act of calling routeQuery causes the
mxs::Reply state of the backend DCB to be modified. It should be
asserted that no call to routeQuery is made while a clientReply call is
in process.

Currently this was only done by readwritesplit when it routed the
queries that were queued up for later execution. The call can be safely
done at a later point as all queries that are received before the lcall
resolves will be appended to the query queue.