Use newer compilers with SLES
Try to install all versions of GCC that have been released. This makes
it automatically use a newer compiler version if a new service pack is
released.

Due to a bug in GCC in SLES 15, the check for whether ASAN works with
crypt_r cannot rely on only the GCC version as on some platforms it
works with GCC 13 and on some it doesn't. To check whether it truly
works requires running a small test program that is compiled with ASAN.