MXS-4962 Initialize vite scaffold
Due to several breaking changes in Vue 3, Vuetify and other libraries,
applying migration to the existing `maxgui` directory is not a good
approach, as the GUI won't be built until all components are migrated.
Regardless, it's recommended to use Vite instead of Webpack for Vue 3.
Therefore, this creates a Vite scaffold in a new directory called
`maxgui_vue3`.