MXS-4967: Acquire timestamp after acquiring lock
This way the current timestamp is always guaranteed to be greater than
or equal to the one stored in the log throttling data. Added assertions
to the code that make sure no underflow ever occurs and increased the
number of threads in the test to add more contention on the mutex.