MXS-4944: Discard config when in static config state
On startup, each object that was defined in a static configuration file
will compute its configuration state and store it. Later on, when
modifications are done at runtime, the current configuration state of
the object is compared to this stored state and if they are identical,
the persisted configuration is deleted as it is now again redundant.

Extended maxctrl_basic with a test case that verifies the runtime
generated file is removed when the configuration goes back into the
original state.