Fix debug assertion in delay_routing()
The assertion would fire if a delay_routing() call with zero delay was
made in the same epoll_wait() batch that returned an error event that
closed the client DCB. Since the reference is still open, the call
should be made even in this case to make sure the delivery always
happens when the reference is alive.

This was triggered by an extra lcall() from readwritesplit that
should've first checked whether there are any queued queries to be
routed before actually doing it.