MXS-4971 Add `TreeTable` component
The `DataTable` was used for most of pages in maxgui to render
tree table and also rowspan feature. It was built on top of
vuetify `v-data-table`. In vuetify 3, it's impossible to reuse
or adjust `DataTable` to adapt to vuetify and vuejs breaking changes.
Therefore, creating a new component that solely supports certain features
would be a wiser choice in terms of maintainability.

This also removes some helper functions as it's now obsolete since
maxscale has required duration type to be a string with specific
unit.