MXS-4971: Remove `devel` package
Hasn't been built for a while and it's far simpler to just build
MaxScale from source to build a custom module. Also removed the
`cdc-connector` package as it also has not been build.

The `all` package is not used and is the same as listing all of the
components in TARGET_COMPONENT.