Don't sleep in tests
A few functions in the test code had excessively long sleeps in them
that were nearly always too long. By replacing them with increasingly
long sleeps or condition variables, the testing speeds up on average
about 0.5s on a local setup. With 418 tests, this results in a test time
reduction of about 200 seconds.