MXS-4971: Remove TARGET_COMPONENT
The TARGET_COMPONENT was a manual workaround for legacy CMake versions
where the component installation did not work and the package names
could not be modified. Since MaxScale now requires CMake 3.16 or newer,
this workaround is no longer needed and the component installation can
be used instead.

Repurposed the old (broken) INSTALL_EXPERIMENTAL variable to control
whether the experimental package is installed or not. Since it's the
only actual package with any content, it serves as a good test case for
the component installation.

Added the installation of the lsb-release package to
install_build_deps.sh since it's not installed by default on
Debian/Ubuntu Docker images. This makes it possible to do packaging with
a docker image which is mostly useful for testing modifications to the
packaging.