MXS-4871 Fix blank page in production build
The issue originates from the `vite-plugin-top-level-await` plugin
which inadvertently removes vuex-persist and some other third-party
modules during the rolling up proccess in the production build.

This removes the plugin as it doesn't provide much benefit for the
current usage, which is to allow export and async import in one-line
code.

This also bumps other packages to latest versions and adds support
for legacy browsers.