MXS-4871 Use attribute inheritance accurately
In vue 3, attribute inheritance has been improved, there is no need
to specify the root element to inheritance the "fallthrough attribute"
as this has been now been done automatically. So, `v-bind="$attrs"`
is no longer needed unless the root element is not the one to inherit it.

This also does some minor chores, cleans up.