MXS-4970: Parse mixed global and local variables
The classification of the SET statement is now fully determined based on
the variables being passed to it. This way, a `SET GLOBAL var=1`
statement gets classified as TYPE_GSYSVAR_WRITE, a `SET var=1` is
classified as TYPE_SESSION_WRITE and `SET GLOBAL var=1, var2=2` is
classified as TYPE_GSYSVAR_WRITE|TYPE_SESSION_WRITE.

The old comment about Oracle's "var := 1" not calling maxscaleKeyword()
being the reason why maxscaleSet() assigns TYPE_SESSION_WRITE is not
correct as maxscaleOracleAssignment() already assigns the type
mask. This is likely a remnant of some old code and the code works as
expected without it.

Added test cases that reproduce the original problem that are also used
to verify that the new behavior works as expected.