MXS-4970: Make pp_mysqlembedded behave like pp_sqlite
The handling of SET statements in pp_mysqlembedded is now closer to that
of pp_sqlite.

There were a handful of test cases where the parsers did not produce the
same result and in all of the cases this was due to pp_mysqlembedded
refusing to parse system variables. Mostly these seem to relate to
replication and to the network.