Make PsTracker::get_args() work with COM_QUERY
This way the "prepared statement" and the arguments for it can be
extracted with the same code regardless of the protocol type.

Changed the return type to std::string as the stored value anyways has
to be copied if it needs to be used later on. Better to do in inside
get_args() as it allows the canonical for the COM_QUERY to be returned.