Create PID directory before checking it
If the PID directory is configured to /var/run/ and the system is
restarted, SystemD would normally create it for us as it's defined in
the SystemD service file. However, if MaxScale is started from the
command line, the directory would not be created as the check for its
existence was done before the code that creates it was run.