MXS-4871 Add `ServiceView`
This also installs `sortablejs` library for reordering
filters in `RelationshipTable`.

In maxgui vue 2, the `RoutingTargetTable` component is used for
rendering routing target relationships but the structure of the
component looks pretty much the same as the `RelationshipTable`
component. Therefore, the `RelationshipTable` component is now
extended to support also routing targets.