Use different --piddir in run_ctrl_c.sh
With the change in how the PID directory is created by SystemD, the
/var/run/maxscale directory no longer remains after the maxscale service
has stopped. The directory would be created automatically by MaxScale
but the maxscale user needs write permission into the directory where
the PID directory would be created.

To make the testing simpler, use --piddir=/tmp in the test.