Wait for monitor in mxs4092_distinct_tables
This will make sure that the shard mapping will not be interrupted by
the loss of the Slave status. This also reveals that there's a built-in
assumption in mariadbmon that a loss of the Slave status is to be always
considered a fatal error. For the schemarouter, using galeramon would be
a better alternative.