MXS-4871 Fix the `Advanced` switch feature of `ParametersTable` component
Toggling the switch should not erase the data the users have typed in.
This was done properly in the old version but after the migration, it
was broken.

The `paramInfoMap` will now have the `hidden` property so that the
`TreeTable` component can use that property to conditionally show
table's `td` element.