MXS-4871 Fix `allowEnterToSubmit` feature of `BaseDlg`
`VTextarea` component from Vuetify version 3 no longer stop event
propagation so it must be stopped manually.

This uses `tabindex` attribute to make dialog focusable so that
keydown event can be listened.

This also renames VForm's modelValue name to `formValidity` to better
reflect that it includes the possibility of being `null`.