MXS-4871 Remove custom button font-size
Overriding button size is a legacy approach that shouldn't be used.
Using Vuetify's default size is a wiser approach, as they set the
default button font-size to 0.875rem, which is 14px. Most of the
buttons' font size in the GUI are either 14px or 12px and Vuetify
supports both size already.

This also updates vuetify to the latest version to fix an issue
with `VDatePicker`, which is the component used by the `LogsArchive`
view.
https://github.com/vuetifyjs/vuetify/issues/19202