MXS-4871 Configure `VTab` style by using Vuetify helper classes
The tab components in the workspace view have slight differences in
styles, so the global customized style must be adjustable.

This declares the global styles by using Vuetify helper classes,
which makes it easier to customize the style of the tab component
without relying on the `!important` rule, thus allowing for further
customization.