MXS-4871 Add `QueryTabNavCtr` and its children components
The `saveFile` mixins has been replaced with `useSaveFile`
composable hook as mixin usage is no longer supported in vue 3.

This also handles the case where the active tab is deleted,
the new active tab must be selected manually as Vuetify has removed
the ability to automatically choose the last tab as the active one.