MXS-4871 Adding back the `closeImmediate` props to `BaseDlg`
In hindsight, this prop is needed for the dialog used in the
`TxtEditorToolbarCtr` component. After confirming the query
execution the dialog should close immediately so that the users can
perform other tasks, such as stopping queries or creating another
query tab, while waiting for the result.

This also sets title's white-space to `normal` as vuetify has removed
this style for `VCardTitle` component, which causes long title text
to not break into a new line.